+++ lib/sector.dart
@@ -92,6 +92,7 @@ export 'src/grids/grid.dart' show EfficientIndexGrid, Grid;
 export 'src/grids/list.dart' show ListGrid;
 export 'src/grids/splay_tree.dart' show SplayTreeGrid;
 export 'src/traverse/draw_line.dart' show drawLine;
+export 'src/traverse/draw_rect.dart' show drawRect;
 export 'src/traverse/row_major.dart' show rowMajor;
 export 'src/utils/grid_extension.dart' show GridExtension;
 export 'src/utils/grid_impl.dart' show GridImpl;